fourth commit

+++ q4.c
@@ -3,7 +3,7 @@
 int main()
 
 {
-    int x=2,y=3,z=3,k=1;
-    printf("%d",3*x/y-z+k);
+    float x=2,y=3,z=3,k=1;
+    printf("%f",3*x/y-z+k);
     return 0;
 }
